Extract utility functions, clean up result (remove 0 values)
