+++ src/pages/AddressInfoPage/AddressTransactionRow.tsx
@@ -140,7 +140,7 @@ const AddressTransactionRow = ({ transaction: t, addressHash, isInContract }: Ad
                 assetId={id}
                 value={amount}
                 highlight={!isMoved}
-                displayAmountSign={!isMoved}
+                displaySign={!isMoved}
                 color={isMoved ? theme.font.secondary : undefined}
                 suffix={symbol}
                 decimals={decimals}
